Исправление теста контроллера, исправление по комментариям

+++ src/main/java/faang/school/projectservice/controller/stage/StageController.java
@@ -12,48 +12,47 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/stage")
+@RequestMapping("/stages")
 @RequiredArgsConstructor
 public class StageController {
     private final StageService stageService;
 
-    @PostMapping("/create")
+    @PostMapping
     public StageDto create(@RequestBody @Valid StageDto stageDto) {
         return stageService.create(stageDto);
     }
 
-    @GetMapping("/get")
+    @GetMapping
     public List<StageDto> findByStatus(@RequestBody ProjectStatus status) {
         return stageService.getByStatus(status);
     }
 
-    @DeleteMapping("/delete/{stageId}")
-    public StageDto deleteStageById(@PathVariable Long stageId) {
+    @DeleteMapping("/{stageId}")
+    public StageDto deleteStageById(@PathVariable Long stageId, @RequestParam(required = false) boolean closeTasks) {
+        if (closeTasks) {
+            return stageService.deleteStageWithClosingTasks(stageId);
+        }
         return stageService.deleteStageById(stageId);
     }
 
-    @DeleteMapping("/delete/{stageId}/close-tasks")
-    public StageDto deleteStageWithClosingTasks(@PathVariable Long stageId) {
-        return stageService.deleteStageWithClosingTasks(stageId);
-    }
-
-    @PutMapping("/update/{stageId}")
+    @PutMapping("/{stageId}")
     public StageDto update(@PathVariable Long stageId, @RequestBody @Valid StageDto stageDto) {
         return stageService.update(stageId, stageDto);
     }
 
-    @GetMapping
+    @GetMapping("/all")
     public List<StageDto> getAll() {
         return stageService.getAll();
     }
 
     @GetMapping("/{stageId}")
-    public StageDto getById(@PathVariable long stageId) {
+    public StageDto getById(@PathVariable Long stageId) {
         return stageService.getById(stageId);
     }
 }
